requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
python-dotenv
fastapi
uvicorn[standard]
pydantic
